integrate AI-Recommendation into src
