--- docker/init-data/AI/book_recommendation_ai.egg-info/PKG-INFO
@@ -1,35 +1,0 @@
-Metadata-Version: 2.4
-Name: book-recommendation-ai
-Version: 0.1.0
-Summary: A FastAPI application for book recommendations using sentence embeddings and FAISS.
-Home-page: https://github.com/yourusername/book-recommendation-ai
-Author: Your Name
-Author-email: your.email@example.com
-Classifier: Programming Language :: Python :: 3
-Classifier: License :: OSI Approved :: MIT License
-Classifier: Operating System :: OS Independent
-Classifier: Topic :: Scientific/Engineering :: Artificial Intelligence
-Classifier: Topic :: Text Processing :: Linguistic
-Classifier: Framework :: FastAPI
-Requires-Python: >=3.9
-Description-Content-Type: text/markdown
-Requires-Dist: pandas>=2.0
-Requires-Dist: fastapi>=0.100
-Requires-Dist: uvicorn[standard]>=0.20
-Requires-Dist: sentence-transformers>=2.2.2
-Requires-Dist: torch>=1.13.0
-Requires-Dist: pyarrow>=10.0.0
-Requires-Dist: fastparquet>=2023.4.0
-Requires-Dist: faiss-cpu>=1.7.0
-Provides-Extra: dev
-Requires-Dist: pytest>=7.0; extra == "dev"
-Requires-Dist: flake8>=6.0; extra == "dev"
-Dynamic: author
-Dynamic: author-email
-Dynamic: classifier
-Dynamic: description-content-type
-Dynamic: home-page
-Dynamic: provides-extra
-Dynamic: requires-dist
-Dynamic: requires-python
-Dynamic: summary
