--- docker/init-data/AI/book_recommendation_ai.egg-info/requires.txt
@@ -1,12 +1,0 @@
-pandas>=2.0
-fastapi>=0.100
-uvicorn[standard]>=0.20
-sentence-transformers>=2.2.2
-torch>=1.13.0
-pyarrow>=10.0.0
-fastparquet>=2023.4.0
-faiss-cpu>=1.7.0
-
-[dev]
-pytest>=7.0
-flake8>=6.0
