--- docker/init-data/AI/book_recommendation_ai.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-
